Add ProductController

+++ FinalProjectDominos/src/main/java/ittalents/dominos/model/entities/PizzaSize.java
@@ -11,10 +11,9 @@ public class PizzaSize {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-
     @Column(name = "name")
     private String name;
-
     @Column(name = "price")
-    private BigDecimal price;
+    private double price;
+
 }
